requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867338309/work
addict==2.4.0
aiohttp @ file:///C:/ci/aiohttp_1632749658027/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1625831764494/work
chardet @ file:///C:/ci/chardet_1607706910910/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
colorama==0.4.4
cryptography @ file:///C:/ci/cryptography_1633502396740/work
cycler==0.10.0
decorator==5.1.0
et-xmlfile==1.1.0
fonttools==4.25.0
future==0.17.1
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///tmp/build/80754af9/gitpython_1624609474443/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
grpcio @ file:///C:/ci/grpcio_1613114370981/work
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
imageio==2.8.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916842152/work
ipdb==0.13.9
ipython==7.29.0
ipywidgets==7.4.2
jedi==0.18.0
jupyter-console==6.0.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
Markdown @ file:///C:/ci/markdown_1614364005059/work
matplotlib~=3.4.3
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mmcv==1.3.16
multidict @ file:///C:/ci/multidict_1607367865052/work
munkres==1.1.4
nbdime==1.0.5
numexpr @ file:///C:/ci/numexpr_1618856761305/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873805/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
opencv-python~=4.5.4.58
openpyxl @ file:///tmp/build/80754af9/openpyxl_1632777717936/work
packaging==21.2
pandas @ file:///C:/ci/pandas_1632920019983/work
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
protobuf==3.17.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.10.0
PyJWT @ file:///C:/ci/pyjwt_1619651810943/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy~=1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035239744/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
termcolor==1.1.0
timm==0.3.2
toml==0.10.2
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets==5.1.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
yacs==0.1.8
yapf==0.31.0
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work

numpy~=1.21.2
blas~=1.0
tqdm~=4.62.3
setuptools~=58.0.4
seaborn~=0.11.2
tensorboardx~=2.2
pandas~=1.3.3